requirements.txt
arrow==0.15.5
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.4
django-allauth==0.41.0
django-autocomplete-light==3.5.1
django-comments-xtd==2.5.1
django-contrib-comments==1.9.2
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-disqus==0.5
django-extensions==2.2.9
django-fluent-comments==2.1
django-import-export==2.1.0
django-markdown2==0.3.1
django-modeltranslation==0.15
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-simple-history==2.10.0
django-sql-explorer==1.1.3
django-tag-parser==3.2
django-taggit==1.2.0
djangorestframework==3.11.0
docutils==0.16
easy-thumbnails==2.7
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.9
jdcal==1.4.1
markdown2==2.3.8
MarkupPy==1.14
mock==4.0.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
psycopg2-binary==2.8.5
python-akismet==0.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
unicodecsv==0.14.1
urllib3==1.25.9
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
